requirements.txt
Django==4.2.6
djangorestframework==3.14.0
Pillow==10.1.0
pymysql==1.1.0